frozen-requirements.txt
appdirs==1.4.4
click==8.1.7
dbus-python==1.2.18
pgi==0.0.11.2
pycairo==1.25.0
PyGObject==3.46.0
